# Cache all trials in Study with delayed relative sampling