--- a/optuna/integration/_lightgbm_tuner/optimize.py
+++ b/optuna/integration/_lightgbm_tuner/optimize.py
@@ -687,7 +687,6 @@ class _LightGBMBaseTuner(_BaseTuner):
                 deepcopy: bool = True,
                 states: Optional[Container[TrialState]] = None,
             ) -> List[optuna.trial.FrozenTrial]:
-
                 trials = super()._get_trials(deepcopy=deepcopy, states=states)
                 return [t for t in trials if t.system_attrs.get(_STEP_NAME_KEY) == self._step_name]