--- a/optuna/integration/_lightgbm_tuner/optimize.py
+++ b/optuna/integration/_lightgbm_tuner/optimize.py
@@ -689,7 +689,7 @@ class _LightGBMBaseTuner(_BaseTuner):
                 use_cache: bool = False,
             ) -> List[optuna.trial.FrozenTrial]:
 
-                trials = super().get_trials(deepcopy=deepcopy, states=states, use_cache=use_cache)
+                trials = super()._get_trials(deepcopy=deepcopy, states=states, use_cache=use_cache)
                 return [t for t in trials if t.system_attrs.get(_STEP_NAME_KEY) == self._step_name]
 
             @property