--- a/optuna/integration/skopt.py
+++ b/optuna/integration/skopt.py
@@ -215,7 +215,7 @@ class SkoptSampler(BaseSampler):
 
     def _get_trials(self, study: Study) -> List[FrozenTrial]:
         complete_trials = []
-        for t in study.get_trials(deepcopy=False):
+        for t in study._get_trials(deepcopy=False, use_cache=True):
             if t.state == TrialState.COMPLETE:
                 complete_trials.append(t)
             elif (