--- a/optuna/pruners/_hyperband.py
+++ b/optuna/pruners/_hyperband.py
@@ -265,7 +265,6 @@ class HyperbandPruner(BasePruner):
             _VALID_ATTRS = (
                 "get_trials",
                 "_get_trials",
-                "_fetch_trials",
                 "directions",
                 "direction",
                 "_directions",