--- a/optuna/samplers/_base.py
+++ b/optuna/samplers/_base.py
@@ -34,6 +34,8 @@ class BaseSampler(abc.ABC):
     is called to determine the relative search space for the trial. Then,
     :meth:`~optuna.samplers.BaseSampler.sample_relative` is invoked to sample parameters
     from the relative search space. During the execution of the objective function,
+    :meth:`~optuna.samplers.BaseSampler.sample_relative` is called only once
+    when sampling the parameters belonging to the relative search space for the first time.
     :meth:`~optuna.samplers.BaseSampler.sample_independent` is used to sample
     parameters that don't belong to the relative search space.