--- a/optuna/study/_optimize.py
+++ b/optuna/study/_optimize.py
@@ -220,8 +220,6 @@ def _run_trial(
         frozen_trial = study._storage.get_trial(trial._trial_id)
         raise
     finally:
-        # We must invalidate all trials cache here as it is only valid within a trial.
-        study._thread_local.cached_all_trials = None
         if frozen_trial.state == TrialState.COMPLETE:
             study._log_completed_trial(frozen_trial)
         elif frozen_trial.state == TrialState.PRUNED: