--- a/optuna/trial/_trial.py
+++ b/optuna/trial/_trial.py
@@ -52,9 +52,6 @@ class Trial(BaseTrial):
         self._study_id = self.study._study_id
         self.storage = self.study._storage
 
-        self._init_relative_params()
-
-    def _init_relative_params(self) -> None:
         self._cached_frozen_trial = self.storage.get_trial(self._trial_id)
         study = pruners._filter_study(self.study, self._cached_frozen_trial)