… tests/integration_tests/test_skopt.py
… tests/samplers_tests/test_cmaes.py
… tests/samplers_tests/test_samplers.py
… tests/samplers_tests/tpe_tests/test_multi_objective_sampler.py
… tests/samplers_tests/tpe_tests/test_sampler.py