@c-bata 2023-03-17
Could you add the following test scenario?

```python
@pytest.mark.parametrize("n_jobs", [1, 2])
def test_cache_is_invalidated(n_jobs) -> None:
    study = create_study()

    def objective(trial: Trial) -> float:
        assert trial._relative_params is None
        assert study._thread_local.cached_all_trials is None

        trial.suggest_float("x", -10, 10)
        assert trial._relative_params is not None
        assert study._thread_local.cached_all_trials is not None
        return -1

    study.optimize(objective, n_trials=10, n_jobs=n_jobs)
```